revert the merge

+++ DoubleFile/DoubleFile/UtilAnalysis_DirList Local/Tree/DetailsDatum/NodeDatum.cs
@@ -21,7 +21,7 @@ namespace Local
         {
             get
             {
-                return new FolderKeyStruct((ulong)nTotalLength, nFilesInSubdirs, nDirsWithFiles);
+                return new FolderKeyStruct(nTotalLength, nFilesInSubdirs, nDirsWithFiles);
             }
         }
